Introduce paragraphGroups and rich paragraphs.
The former 'paragraph' returned an array which was because some nested
NLM paragraph elements are represented in a flat way in the Substance based
document format.
To better reflect this, I renamed the method.
The actual paragraph is now handled by 'richParagraph' which also
allows inline elements such as 'inline-graphic' or 'inline-formula'.